# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elite Fitness Gym API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  member: "member",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  phone?: string;
  role: UserRole;
  planId?: number;
  memberSince?: string;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Trainer {
  id: number;
  name: string;
  specialty: string;
  bio: string;
  imageUrl?: string;
  email?: string;
  phone?: string;
  experience?: number;
  certifications?: string[];
  rating?: number;
  createdAt?: string;
}

export interface CreateTrainerRequest {
  name: string;
  specialty: string;
  bio: string;
  imageUrl?: string;
  email?: string;
  phone?: string;
  experience?: number;
  certifications?: string[];
  rating?: number;
}

export interface BlogPost {
  id: number;
  title: string;
  excerpt: string;
  content: string;
  imageUrl?: string;
  category?: string;
  author: string;
  published?: boolean;
  createdAt?: string;
}

export interface BlogPostsResponse {
  posts: BlogPost[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateBlogPostRequest {
  title: string;
  excerpt: string;
  content: string;
  imageUrl?: string;
  category?: string;
  author: string;
  published?: boolean;
}

export interface Testimonial {
  id: number;
  name: string;
  role?: string;
  content: string;
  rating: number;
  imageUrl?: string;
  approved?: boolean;
  createdAt?: string;
}

export interface CreateTestimonialRequest {
  name: string;
  role?: string;
  content: string;
  rating: number;
  imageUrl?: string;
}

export interface PricingPlan {
  id: number;
  name: string;
  price: number;
  period: string;
  description?: string;
  features: string[];
  highlighted?: boolean;
  color?: string;
}

export interface GalleryImage {
  id: number;
  url: string;
  title: string;
  category?: string;
  createdAt?: string;
}

export interface CreateGalleryImageRequest {
  url: string;
  title: string;
  category?: string;
}

export interface PerformanceEntry {
  id: number;
  userId: number;
  date: string;
  weight: number;
  bmi?: number;
  bodyFat?: number;
  muscleMass?: number;
  notes?: string;
  createdAt?: string;
}

export interface CreatePerformanceEntryRequest {
  date: string;
  weight: number;
  bmi?: number;
  bodyFat?: number;
  muscleMass?: number;
  notes?: string;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  pending: "pending",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface Payment {
  id: number;
  userId: number;
  amount: number;
  dueDate: string;
  paidDate?: string;
  status: PaymentStatus;
  planName?: string;
  description?: string;
}

export interface ContactMessageRequest {
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  phone?: string;
  subject?: string;
  message: string;
  trainerId?: number;
  read?: boolean;
  createdAt?: string;
}

export interface AdminStats {
  totalMembers: number;
  activeMembers: number;
  totalRevenue: number;
  pendingPayments: number;
  totalTrainers: number;
  totalBlogPosts: number;
  unreadMessages: number;
  newMembersThisMonth: number;
}

export type GetBlogPostsParams = {
  page?: number;
  limit?: number;
};