# lib/api-zod/src/generated/types/createGalleryImageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elite Fitness Gym API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGalleryImageRequest {
  url: string;
  title: string;
  category?: string;
}