# lib/api-zod/src/generated/types/pricingPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elite Fitness Gym API
 * OpenAPI spec version: 0.1.0
 */

export interface PricingPlan {
  id: number;
  name: string;
  price: number;
  period: string;
  description?: string;
  features: string[];
  highlighted?: boolean;
  color?: string;
}